extract rpc to module
