+++ ptychodus/api/settings.py
@@ -199,5 +199,5 @@ class SettingsRegistry(Observable):
                                 + valueString[len(self._replacementPathPrefix):]
                 config.set(settingsGroup.name, settingsEntry.name, valueString)
 
-        with open(filePath, 'w') as configFile:
+        with filePath.open(mode='w') as configFile:
             config.write(configFile)
