+++ ptychodus/plugins/neXus/lynxScanFile.py
@@ -40,7 +40,7 @@ class LynxScanFileReader(ScanFileReader):
         pointDict: dict[int, list[LynxScanPoint]] = defaultdict(list[LynxScanPoint])
         pointList: list[ScanPoint] = list()
 
-        with open(filePath, newline='') as csvFile:
+        with filePath.open(newline='') as csvFile:
             csvReader = csv.reader(csvFile, delimiter=' ')
             csvIterator = iter(csvReader)
